Turns out new WebGL algorithm does good even without gradual resizing. So drop it for now.
