+++ src/demo/html/demo.html
@@ -46,10 +46,6 @@
       <form id="resize" class="options">
         w: <input type="text" name="width" />
         h: <input type="text" name="height" />
-        gradient: <select name="gradient">
-          <option value="1">use (better quality)</option>
-          <option value="0">do not use</option>
-        </select>
         <button type="submit">Go</button>
       </form>
